Add a new version of Caprice
The main code of Caprice is by Ulrich Doewich.
Ramlaid/Arkos wrote a C++ version of caprice with some rewriting for CRTC and screen gestion.
Krusty/Benediction applied some modification in order to use the libdsk and real diskc (not yet tested) and has begun to write a wxWidget interface.